refactor: SSE이벤트 발행 비동기 처리로 변경
