+++ backend/src/main/java/team/teamby/teambyteam/global/configuration/EnableAsyncConfiguration.java
@@ -1,9 +1,9 @@
-package team.teamby.teambyteam.icalendar.cofig;
+package team.teamby.teambyteam.global.configuration;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.scheduling.annotation.EnableAsync;
 
 @Configuration
 @EnableAsync
-public class IcalendarConfiguration {
+public class EnableAsyncConfiguration {
 }
